winutils/utility_manager/requirements.txt
win11toast==0.32
pynput==1.7.6
pywin32==306
platformdirs==3.2.0
psutil==5.9.5
pystray==0.19.4
Pillow==10.0.0
